+++ tests/unit/test_charm.py
@@ -16,7 +16,7 @@ import pytest
 import yaml
 from github_runner_manager import constants
 from github_runner_manager.platform.platform_provider import TokenError
-from ops.model import BlockedStatus, StatusBase, WaitingStatus
+from ops.model import BlockedStatus, StatusBase
 from ops.testing import Harness
 
 from charm import (
@@ -30,7 +30,6 @@ from charm_state import (
     FLAVOR_LABEL_COMBINATIONS_CONFIG_NAME,
     IMAGE_INTEGRATION_NAME,
     LABELS_CONFIG_NAME,
-    MONGO_DB_INTEGRATION_NAME,
     OPENSTACK_CLOUDS_YAML_CONFIG_NAME,
     OPENSTACK_FLAVOR_CONFIG_NAME,
     PATH_CONFIG_NAME,
@@ -52,7 +51,6 @@ from errors import (
     ImageIntegrationMissingError,
     ImageNotFoundError,
     LogrotateSetupError,
-    MissingMongoDBError,
     RunnerError,
     SubprocessError,
 )
@@ -363,22 +361,6 @@ def test_on_install_failure(
     assert "mock stderr" in str(exc.value)
 
 
-def test_charm_goes_into_waiting_state_on_missing_integration_data(
-    monkeypatch: pytest.MonkeyPatch, harness: Harness
-):
-    """
-    arrange: Mock charm._setup_state to raise an MissingIntegrationDataError.
-    act: Fire config changed event.
-    assert: Charm is in blocked state.
-    """
-    setup_state_mock = MagicMock(side_effect=MissingMongoDBError("mock error"))
-    monkeypatch.setattr(GithubRunnerCharm, "_setup_state", setup_state_mock)
-    harness.update_config({PATH_CONFIG_NAME: "mockorg/repo", TOKEN_CONFIG_NAME: "mocktoken"})
-    harness.charm.on.config_changed.emit()
-    assert isinstance(harness.charm.unit.status, WaitingStatus)
-    assert "mock error" in harness.charm.unit.status.message
-
-
 def test_check_runners_action_with_errors():
     mock_event = MagicMock()
 
@@ -683,34 +665,6 @@ def test_attempting_disable_legacy_service_for_upgrade(
         [mock.call(LEGACY_RECONCILE_TIMER_SERVICE), mock.call(LEGACY_RECONCILE_SERVICE)],
         any_order=True,
     )
-
-
-@pytest.mark.parametrize(
-    "hook",
-    [
-        pytest.param("database_created", id="Database Created"),
-        pytest.param("endpoints_changed", id="Endpoints Changed"),
-        pytest.param("mongodb_relation_broken", id="MongoDB Relation Departed"),
-    ],
-)
-def test_database_integration_events_setup_service(
-    hook: str, monkeypatch: pytest.MonkeyPatch, harness: Harness
-):
-    """
-    arrange: Mock charm._setup_service.
-    act: Fire mongodb relation events.
-    assert: _setup_service has been called.
-    """
-    setup_service_mock = MagicMock()
-    relation_mock = MagicMock()
-    relation_mock.name = "mongodb"
-    relation_mock.id = 0
-    monkeypatch.setattr("charm.GithubRunnerCharm._setup_service", setup_service_mock)
-    if hook.startswith(MONGO_DB_INTEGRATION_NAME):
-        getattr(harness.charm.on, hook).emit(relation=relation_mock)
-    else:
-        getattr(harness.charm.database.on, hook).emit(relation=relation_mock)
-    setup_service_mock.assert_called_once()
 
 
 def test_planner_relation_changed_writes_flavor(monkeypatch: pytest.MonkeyPatch):